requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
discord==1.0.1
discord.py==1.0.1
Flask==1.0.3
gevent==1.4.0
greenlet==0.4.15
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
requests==2.21.0
six==1.12.0
typing-extensions==3.7.2
urllib3==1.24.3
waitress==1.3.0
websocket-client==0.56.0
websockets==6.0
Werkzeug==0.15.4
yarl==1.3.0
